api/requirements.txt
flask==2.0.1
flask-cors==3.0.10
pyodbc==4.0.32
python-dotenv